just: fix cross-compilation
Get rid of calls to `cargo run` in a cross-compilation stdenv, which
lead to cargo TARGET and CC to be out of sync.

Use emulator to build man pages and shell completions.

Build doc as a separate derivation with buildPackages, because that
would be prohibitively expensive to do via emulation.